% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Gene_Enrichment.R
\name{Run_Gene_Enrichment}
\alias{Run_Gene_Enrichment}
\title{Title  Gene enrichment analysis ORA}
\usage{
Run_Gene_Enrichment(geneList = geneList, org = "hsa")
}
\arguments{
\item{geneList}{geneList is which gene list you want to ORA analysis}

\item{org}{The spesis names. defalt = "hsa"}
}
\value{
null
}
\description{
This function is used to human or mouse gene ORA analysis
}
\examples{
 Run_Gene_Enrichment(geneList = genelist,org = "hsa")
}
